Modified play loop to play very quick notes more accurately
The play loop now accumulates all waits until there is a wait time of at least 0.016s (one game tick) before playing a note/chord. It then goes through the list of bots and assigns each of them a note. The Read note and Play note rules have been combined into one to avoid repetition.

There is now a race condition in the Play note rule call (which has been done with a boolean, and can't be modified to be a subroutine due to limitations of the workshop). The script contains a workaround for this, resetting a bot's playNote variable if it has been True for over 0.200 seconds.